requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Bootstrap4==4.0.2
Flask-Mail==0.9.1
numpy==1.24.3
matplotlib==3.7.2
python-dotenv==1.0.0
email-validator==2.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.20 